refactor: change test signature and fields
