+++ src/test/java/qna/domain/QuestionTest.java
@@ -1,28 +1,33 @@
 package qna.domain;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
-import java.util.List;
-
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import qna.CannotDeleteException;
 
 public class QuestionTest {
-    public static final Question Q1 = new Question("title1", "contents1", UserTest.JAVAJIGI);
-    public static final Question Q2 = new Question("title2", "contents2", UserTest.SANJIGI);
 
-    @DisplayName("질문을 삭제하고 삭제 히스토리를 반환한다.")
+    public Question question;
+    public User guestUser;
+
+    @BeforeEach
+    void setUp() {
+        question = new Question("How to learn java?", "I want to be a java master", UserTest.JAVAJIGI);
+        Answer A1 = new Answer(UserTest.JAVAJIGI, question, "Answers Contents1");
+        Answer A2 = new Answer(UserTest.SANJIGI, question, "Answers Contents2");
+        question.addAnswer(A1);
+        question.addAnswer(A2);
+        guestUser = User.GUEST_USER;
+    }
+
+    @DisplayName("질문을 삭제할 때 답변자가 자기 자신이 아닌 경우 삭제할 수 없다.")
     @Test
-    void 질문을_삭제하고_이를_반환한다() throws CannotDeleteException {
-        List<DeleteHistory> deletedTheFirst = Q1.delete(UserTest.JAVAJIGI);
-        List<DeleteHistory> deletedTheSecond = Q2.delete(UserTest.SANJIGI);
-
-        assertAll(
-            () -> assertThat(deletedTheFirst.get(0)).isEqualTo(DeleteHistory.class),
-            () -> assertThat(deletedTheSecond.get(0)).isEqualTo(DeleteHistory.class)
-        );
+    void 질문_삭제시_답변자가_자신이_아니면_삭제할_수_없다() {
+        assertThatThrownBy(
+            () -> question.delete(guestUser)
+        ).isInstanceOf(CannotDeleteException.class);
     }
 }
